Ported redis-py to python 3
The main/interesting modifications are to do with encoding. All user data is
returned as "bytes", and only redis data (like the output of the INFO) command
is returned as unicode. There may be some inconcistencies here however. There
are no other changes to the API.

Internally, iterd in redis.compat should be use to iterate of dictionaries, and
when comparing string literals to reponces, the strings need to be encoded.